requirements.txt
Django==5.0.6
gunicorn==22.0.0
psycopg[binary,pool]==3.1.19
dj-database-url==2.2.0
djangorestframework==3.15.1
django-filter==24.2
drf-spectacular[sidecar]==0.27.2
pytest-django==4.8.0
dj-rest-auth==6.0.0
djangorestframework-simplejwt==5.3.1
django-allauth[socialaccount]==0.63.3
factory_boy==3.3.0
